Simplify L3 model (no base class)
