+++ country-a-service/cda-generator/src/main/java/dk/sds/ncp/cda/model/EPrescriptionL3.java
@@ -1,21 +1,29 @@
 package dk.sds.ncp.cda.model;
 
 import dk.sds.ncp.cda.Utils;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-import lombok.experimental.SuperBuilder;
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.Value;
 
 import java.time.OffsetDateTime;
 
-@EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
-@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-@Getter
-@SuperBuilder(toBuilder = true)
-@SuppressWarnings("unused")
-public class EPrescriptionL3 extends EPrescriptionBase{
+@Value
+@Builder
+public class EPrescriptionL3 {
+    /**
+     * Unique ID identifying the CDA document (not the prescription itself).
+     */
+    @NonNull CdaId documentId;
+    @NonNull String title;
+    @NonNull CdaId prescriptionId;
 
+    /**
+     * "[...] the date and time at which this document was created as an electronic document."
+     */
+    @NonNull OffsetDateTime effectiveTime;
+    @NonNull Author author;
 
+    @NonNull String indicationText;
 
     @NonNull Patient patient;
     /**
@@ -30,6 +38,12 @@ public class EPrescriptionL3 extends EPrescriptionBase{
     @NonNull Long packageQuantity;
     @NonNull Boolean substitutionAllowed;
 
+    public String getEffectiveTime() {
+        return Utils.cdaDateTime(effectiveTime);
+    }
+    public OffsetDateTime getEffectiveTimeOffsetDateTime() {
+        return effectiveTime;
+    }
     public String getSignatureTime() {
         return Utils.cdaDateTime(signatureTime);
     }
